Create vm-check
The terminal can ask users for a resource group name. If the inputted name does not exist the user can create a new resource group under that name. After assigning the resource group successfully the terminal asks users for a VM name and confirms if it either exists or doesn't.